requirements.txt
absl-py==2.2.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
easyocr==1.7.2
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
jamo==0.4.1
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.6.2
loguru==0.7.3
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.4
numpy==2.0.2
onnx==1.17.0
onnx-simplifier==0.4.10
onnxruntime==1.19.2
onnxsim==0.4.36
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.10.0
paddlepaddle==3.0.0
pillow==11.2.1
protobuf==6.30.2
psutil==7.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.8
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
rich==14.0.0
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
thop==0.1.1.post2209072238
tifffile==2024.8.30
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
-e git+https://github.com/Megvii-BaseDetection/YOLOX.git@d872c71bf63e1906ef7b7bb5a9d7a529c7a59e6a#egg=yolox
zipp==3.21.0
